+++ CHANGELOG.md
@@ -10,9 +10,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 ### Changed
+
+- Improved test function to cover additional CloudEvent cases
+
 ### Deprecated
 ### Removed
 ### Fixed
+
+- CloudEvent handler gracefully fails on non-cloudevent requests
+
 ### Security
 
 ## [0.5.1] - 2025-03-10